Code Refactor and Addition of Feature
- Minor optimisations
- Resolved #13
- Separated the Cobra CMDs into their own files
- Added a .bat file to compile the program

+++ src/spinner/spinner_types.go
@@ -2,7 +2,7 @@
 package spinner
 
 type SpinnerInfo struct {
-	Interval float32
+	Interval int64
 	Frames   []string
 }
